Moving dtype related items to nu_dtype
